[#14] - rename a column in collection
